available_tutorials {learnr}	R Documentation

List available tutorials

Description

List the tutorials that are currently available via installed R packages. Or list the specific tutorials that are contained within a given R package.

Usage

available_tutorials(package = NULL)

Arguments

package
Name of package

Value

available_tutorials() returns a data.frame containing "package", "name", "title", "description", "package_dependencies", "private", and "yaml_front_matter".

Examples

available_tutorials(package = "learnr")

[Package learnr version 0.10.1.9019 Index]