filesystem_storage {learnr}	R Documentation

Filesystem-based storage for tutor state data

Description

Tutorial state storage handler that uses the filesystem as a backing store. The directory will contain tutorial state data partitioned by user_id, tutorial_id, and tutorial_version (in that order)

Usage

filesystem_storage(dir, compress = TRUE)

Arguments

dir
Directory to store state data within

compress
Should .rds files be compressed?

Value

Storage handler suitable for options(tutorial.storage = ...)

[Package learnr version 0.10.1.9019 Index]