finalize_question {learnr}	R Documentation

Finalize a question

Description

Mark a question as finalized by adding a question-final class to the HTML output at the top level, in addition to disabling all tags with disable_all_tags().

Usage

finalize_question(ele)

Arguments

ele
html tag element

Examples

# finalize the question UI
finalize_question(
  htmltools::div(
    class = "custom-question",
    htmltools::div("answer 1"),
    htmltools::div("answer 2")
  )
)

[Package learnr version 0.10.1.9019 Index]