learnr-package {learnr}	R Documentation

learnr: Interactive Tutorials for R

Description

Create interactive tutorials using R Markdown. Use a combination of narrative, figures, videos, exercises, and quizzes to create self-paced tutorials for learning about R and R packages.

Author(s)

Maintainer: Garrick Aden-Buie garrick@rstudio.com (ORCID)

Authors:

Barret Schloerke barret@rstudio.com (ORCID)

JJ Allaire jj@rstudio.com [conceptor]

Other contributors:

Alexander Rossell Hayes alex.rossellhayes@rstudio.com (ORCID) [contributor]

Nischal Shrestha nischal@rstudio.com (ORCID) [contributor]

Angela Li angelali921@gmail.com (vignette) [contributor]

RStudio [copyright holder, funder]

Ajax.org B.V. (Ace library) [contributor, copyright holder]

Zeno Rocha (clipboard.js library) [contributor, copyright holder]

Nick Payne (Bootbox library) [contributor, copyright holder]

Jake Archibald (idb-keyval library) [contributor, copyright holder]

See Also

Useful links:

https://rstudio.github.io/learnr/

https://github.com/rstudio/learnr

Report bugs at https://github.com/rstudio/learnr/issues

[Package learnr version 0.10.1.9019 Index]