random_praise {learnr}	R Documentation

Random praise and encouragement

Description

Random praises and encouragements sayings to compliment your question and quiz experience.

Usage

random_praise(language = NULL)

random_encouragement(language = NULL)

Arguments

language
The language for the random phrase. The currently supported languages include: en, es, pt, tr, de, emo, and testing (static phrases).

Value

Character string with a random saying

[Package learnr version 0.10.1.9019 Index]