safe_env {learnr}	R Documentation

Safe R CMD environment

Description

By default, callr::rcmd_safe_env suppresses the ability to open a browser window. This is the default execution environment within callr::r. However, opening a browser is expected behavior within the learnr package and should not be suppressed.

Usage

safe_env()

[Package learnr version 0.10.1.9019 Index]