Test the exercises of a learnr tutorial

Runs the automated tests for interactive exercises for a tutorial. Interactive exercises may have associated tests that are specified in a <label>-tests chunk, where <label> is the chunk label of the exercise chunk (the chunk with exercise = TRUE). test_tutorial() can be used interactively or within a testthat test suite.

Here's a simple tutorial with a single interactive exercise with several test cases:

`r paste(readLines(system.file("internals", "example-tutorial-test.Rmd", package = "learnr")), collapse = "\n")`

Usage

test_tutorial(
  path,
  test = c("all", "solutions", "tests"),
  ...,
  quiet = TRUE,
  safely = TRUE,
  reporter = NULL
)

Arguments

path

[character()]
The path or a vector of paths to tutorial .Rmd source files.

test

[character(1)]
Which tests should be run? The default is "all", which runs both the stated tests and tests that the exercise solution passes. To only test stated tests in the -tests chunk, use "tests" and to only test the solution, use "solution".

...

Arguments passed on to rmarkdown::render

input

The input file to be rendered. This can be an R script (.R), an R Markdown document (.Rmd), or a plain markdown document.

output_format

The R Markdown output format to convert to. The option "all" will render all formats defined within the file. The option can be the name of a format (e.g. "html_document") and that will render the document to that single format. One can also use a vector of format names to render to multiple formats. Alternatively, you can pass an output format object (e.g. html_document()). If using NULL then the output format is the first one defined in the YAML frontmatter in the input file (this defaults to HTML if no format is specified there). If you pass an output format object to output_format, the options specified in the YAML header or _output.yml will be ignored and you must explicitly set all the options you want when you construct the object. If you pass a string, the output format will use the output parameters in the YAML header or _output.yml.

output_file

The name of the output file. If using NULL then the output filename will be based on filename for the input file. If a filename is provided, a path to the output file can also be provided. Note that the output_dir option allows for specifying the output file path as well, however, if also specifying the path, the directory must exist. If output_file is specified but does not have a file extension, an extension will be automatically added according to the output format. To avoid the automatic file extension, put the output_file value in I(), e.g., I('my-output').

output_dir

The output directory for the rendered output_file. This allows for a choice of an alternate directory to which the output file should be written (the default output directory of that of the input file). If a path is provided with a filename in output_file the directory specified here will take precedence. Please note that any directory path provided will create any necessary directories if they do not exist.

output_options

List of output options that can override the options specified in metadata (e.g. could be used to force self_contained or mathjax = "local"). Note that this is only valid when the output format is read from metadata (i.e. not a custom format object passed to output_format).

output_yaml

Paths to YAML files specifying output formats and their configurations. The first existing one is used. If none are found, then the function searches YAML files specified to the output_yaml top-level parameter in the YAML front matter, _output.yml or _output.yaml, and then uses the first existing one.

intermediates_dir

Intermediate files directory. If a path is specified then intermediate files will be written to that path. If NULL, intermediate files are written to the same directory as the input file.

knit_root_dir

The working directory in which to knit the document; uses knitr's root.dir knit option. If NULL then the behavior will follow the knitr default, which is to use the parent directory of the document.

runtime

The runtime target for rendering. The static option produces output intended for static files; shiny produces output suitable for use in a Shiny document (see run). The default, auto, allows the runtime target specified in the YAML metadata to take precedence, and renders for a static runtime target otherwise.

clean

Using TRUE will clean intermediate files that are created during rendering.

params

A list of named parameters that override custom params specified within the YAML front-matter (e.g. specifying a dataset to read or a date range to confine output to). Pass "ask" to start an application that helps guide parameter configuration.

knit_meta

(This option is reserved for expert use.) Metadata generated by knitr.

envir

The environment in which the code chunks are to be evaluated during knitting (can use new.env() to guarantee an empty new environment).

run_pandoc

An option for whether to run pandoc to convert Markdown output.

encoding

Ignored. The encoding is always assumed to be UTF-8.

quiet

An option to suppress printing during rendering from knitr, pandoc command line and others. To only suppress printing of the last "Output created: " message, you can set rmarkdown.render.message to FALSE

safely

[logical(1)]
If TRUE, test_tutorial() will render and test the tutorial in an isolated (safe) background process using safe(). If FALSE, test_tutorial() will render and test the tutorial in the current R process, which may attach packages or modify the global environment and should only be used when iterating with a single tutorial.

reporter

Reporter to use to summarise output. Can be supplied as a string (e.g. "summary") or as an R6 object (e.g. SummaryReporter$new()).

See Reporter for more details and a list of built-in reporters.

Value

Nothing.

Examples

# This could be included in a package's testthat tests
# e.g. tests/testthat/test-tutorials.R

tutorial_path <- system.file(
  "internals", "example-tutorial-test.Rmd", package = "learnr"
)

if (!requireNamespace("testthat", quietly = TRUE)) {
  test_tutorial(tutorial_path)
}