tutorial_package_dependencies {learnr}	R Documentation

List tutorial dependencies

Description

List the R packages required to run a particular tutorial.

Usage

tutorial_package_dependencies(name = NULL, package = NULL)

Arguments

name
The tutorial name. If name is NULL, then all tutorials within package will be searched.

package
The R package providing the tutorial. If package is NULL, then all tutorials will be searched.

Value

A character vector of package names that are required for execution.

Examples

tutorial_package_dependencies(package = "learnr")

[Package learnr version 0.10.1.9019 Index]