Tutorial HTML dependency

Tutorial HTML dependency

Usage

tutorial_html_dependency()

Details

HTML dependency for core tutorial JS and CSS. This should be included as a dependency for custom tutorial formats that wish to ensure that that tutorial.js and tutorial.css are loaded prior their own scripts and stylesheets.